requirements.txt
annotated-types==0.7.0
anyio==4.6.0
azure-cognitiveservices-speech==1.40.0
blinker==1.7.0
boto3==1.34.36
botocore==1.34.36
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
distro==1.9.0
Flask==3.0.1
Flask-Session==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identity==0.3.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.4
msal==1.26.0
openai==1.13.3
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.0.7
Werkzeug==3.0.1
